75 hrpp

+++ src/materials/dialectric.rs
@@ -1,6 +1,6 @@
 use std::ops::Neg;
 
-use glam::dvec3;
+use glam::vec3;
 use rand::random;
 
 use crate::{hittable::HitRecord, ray::Ray};
@@ -12,18 +12,18 @@ use super::{
 
 #[derive(Clone, Copy)]
 pub struct Dialectric {
-    pub index_of_refraction: f64,
+    pub index_of_refraction: f32,
 }
 
 impl Dialectric {
-    pub fn new(index_of_refraction: f64) -> Dialectric {
+    pub fn new(index_of_refraction: f32) -> Dialectric {
         Dialectric {
             index_of_refraction,
         }
     }
 
     /// Shclick's approximation for reflectance
-    fn reflectance(cos: f64, ref_idx: f64) -> f64 {
+    fn reflectance(cos: f32, ref_idx: f32) -> f32 {
         let r0 = ((1.0 - ref_idx) / (1.0 + ref_idx)).powi(2);
         r0 + (1.0 - r0) * (1.0 - cos).powi(5)
     }
@@ -31,7 +31,7 @@ impl Dialectric {
 
 impl Material for Dialectric {
     fn scatter(&self, ray: &Ray, hit_record: &HitRecord) -> Option<ScatterRecord> {
-        let attenuation = dvec3(1.0, 1.0, 1.0);
+        let attenuation = vec3(1.0, 1.0, 1.0);
         let refraction_ratio = if hit_record.front_face {
             1.0 / self.index_of_refraction
         } else {
@@ -39,13 +39,13 @@ impl Material for Dialectric {
         };
         let unit_direction = ray.direction.normalize();
 
-        let cos_theta = f64::min(unit_direction.neg().dot(hit_record.normal), 1.0);
-        let sin_theta = f64::sqrt(1.0 - cos_theta.powi(2));
+        let cos_theta = f32::min(unit_direction.neg().dot(hit_record.normal), 1.0);
+        let sin_theta = f32::sqrt(1.0 - cos_theta.powi(2));
 
         let cannot_refract = refraction_ratio * sin_theta > 1.0;
 
         let direction = if cannot_refract
-            || Dialectric::reflectance(cos_theta, refraction_ratio) > random::<f64>()
+            || Dialectric::reflectance(cos_theta, refraction_ratio) > random::<f32>()
         {
             utils::reflect(unit_direction, hit_record.normal)
         } else {
